Add App.post_to_slack
* Also require time on app.

Why?
Because Time.strptime is part of ruby standard library and not core
library. https://ruby-doc.org/stdlib-2.6.3/libdoc/time/rdoc/index.html

Why it was working before?
Because app.rb was required only in server.rb and it happens that it was
also required webrick, which internally was requiring time.
https://github.com/ruby/ruby/blob/fd20b32130f52f4a7799a8d7a367f5b39636bc6a/lib/webrick.rb
https://github.com/ruby/ruby/blob/fd20b32130f52f4a7799a8d7a367f5b39636bc6a/lib/webrick/cookie.rb#L12